smartagri/backend/requirements.txt
absl-py==0.7.1
alabaster==0.7.12
astor==0.8.0
astroid==2.2.5
astropy==3.2.1
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.7.0
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.1.0
boto==2.49.0
boto3==1.9.187
botocore==1.12.187
branca==0.3.1
bs4==0.0.1
casttube==0.2.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
confuse==1.0.0
cycler==0.10.0
Cython==0.29.7
decorator==4.4.0
defusedxml==0.6.0
Django==2.1
django-filter==2.2.0
django-redis==4.11.0
django-redis-sessions==0.6.1
djangorestframework==3.11.0
docutils==0.14
drf-writable-nested==0.5.4
echarts-china-cities-pypkg==0.0.9
echarts-china-counties-pypkg==0.0.2
echarts-china-misc-pypkg==0.0.1
echarts-china-provinces-pypkg==0.0.3
echarts-countries-pypkg==0.1.6
entrypoints==0.3
ez-setup==0.9
flake8==3.7.7
Flask==1.1.1
folium==0.10.0
future==0.17.1
gast==0.2.2
gensim==3.8.0
grpcio==1.21.1
gTTS==2.0.4
gTTS-token==1.1.3
h5py==2.9.0
htmlmin==0.1.12
icrawler==0.6.2
idna==2.8
ifaddr==0.1.6
imageio==2.5.0
imagesize==1.1.0
imgaug==0.2.9
importlib-metadata==0.19
ipykernel==5.1.1
ipyparallel==6.2.4
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.20
itsdangerous==1.1.0
Janome==0.3.9
jedi==0.13.3
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
json5==0.8.5
jsonschema==3.0.1
jupyter-client==5.2.4
jupyter-core==4.4.0
jupyterlab==1.2.4
jupyterlab-server==1.0.6
kaggle==1.5.4
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
lineedit==0.1.1
llvmlite==0.29.0
lml==0.0.2
lxml==4.3.3
Markdown==3.1.1
MarkupSafe==1.1.1
mask-rcnn==2.1
matplotlib==3.1.0
mccabe==0.6.1
mecab-python-windows==0.996.3
missingno==0.4.2
mistune==0.8.4
mock==3.0.5
more-itertools==7.2.0
mysqlclient==1.4.4
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.3
nltk==3.4.4
nose==1.3.7
notebook==5.7.8
numba==0.45.1
numpy==1.16.3
opencv-python==4.1.0.25
packaging==19.0
paho-mqtt==1.5.0
pandas==0.24.2
pandas-profiling==2.3.0
pandocfilters==1.4.2
parso==0.4.0
patsy==0.5.1
phik==0.9.8
pickleshare==0.7.5
Pillow==6.0.0
pluggy==0.12.0
prettytable==0.7.2
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.1
py==1.8.0
pybind11==2.3.0
PyChromecast==4.1.0
pycodestyle==2.5.0
pycparser==2.19
pyecharts==0.1.9.4
pyflakes==2.1.1
Pygments==2.4.2
pylint==2.3.1
PyMySQL==0.9.3
pyparsing==2.4.0
PyQt4==4.11.4
PyQt5==5.12.2
PyQt5-sip==4.19.17
pyrsistent==0.15.2
pytest==5.1.1
pytest-pylint==0.14.1
python-dateutil==2.8.0
python-slugify==3.0.2
pytz==2019.1
PyWavelets==1.0.3
pywinpty==0.5.5
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.5.1
radian==0.4.0
rchitect==0.3.4
redis==3.3.11
requests==2.22.0
s3transfer==0.2.1
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.3.0
seaborn==0.9.0
selenium==3.141.0
Send2Trash==1.5.0
Shapely==1.6.4.post1
simplejson==3.16.0
sip==4.19.8
six==1.12.0
sklearn==0.0
smart-open==1.8.4
snowballstemmer==1.2.1
soupsieve==1.9.1
Sphinx==2.0.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
statsmodels==0.10.1
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
text-unidecode==1.2
tornado==6.0.3
tqdm==4.32.1
traitlets==4.3.2
typed-ast==1.3.5
urllib3==1.24.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.4.2
wordcloud==1.5.0
wrapt==1.11.1
xgboost==0.90
xlrd==1.2.0
yapf==0.27.0
zeroconf==0.23.0
zipp==0.6.0
